packages/cli: rename lib/bundle to bundler + serveBundle
