+++ packages/cli/src/lib/bundler/optimization.ts
