+++ packages/cli/src/lib/bundler/paths.ts
